+++ Nos/Views/AuthorCard.swift
@@ -7,108 +7,87 @@
 
 import SwiftUI
 
-/// This view displays the information we have for an Author suitable for being used in a grid.
+/// This view displays the information we have for an author suitable for being used in a list.
 struct AuthorCard: View {
-   
-    @ObservedObject var author: Author 
-    
-    @EnvironmentObject private var router: Router
-    @EnvironmentObject private var currentUser: CurrentUser
-    
-    var followersRequest: FetchRequest<Follow>
-    var followersResult: FetchedResults<Follow> { followersRequest.wrappedValue }
 
-    var followers: Followed {
-        followersResult.map { $0 }
-    }
+    @ObservedObject var author: Author
 
-    private var knownFollowers: [Follow] {
-        author.followers.filter {
-            guard let source = $0.source else {
-                return false
-            }
-            return source.hasHumanFriendlyName == true &&
-                source != author &&
-                (currentUser.isFollowing(author: source) || currentUser.isBeingFollowedBy(author: source))
-        }
-    }
-    
-    init(author: Author) {
+    @Environment(\.managedObjectContext) private var viewContext
+
+    var tapAction: (() -> Void)?
+ 
+    init(author: Author, onTap: (() -> Void)? = nil) {
         self.author = author
-        self.followersRequest = FetchRequest(fetchRequest: Follow.followsRequest(destination: [author]))
+        self.tapAction = onTap
     }
-    
+
     var body: some View {
         Button {
-            router.push(author)
+            tapAction?()
         } label: {
-            VStack(alignment: .center, spacing: 0) {
-                if let profilePhotoURL = author.profilePhotoURL {
-                    SquareImage(url: profilePhotoURL)
-                        .aspectRatio(contentMode: .fit)
-                        .layoutPriority(1)
-                } else {
-                    Image.emptyAvatar
-                        .resizable()
-                        .renderingMode(.original)
-                        .aspectRatio(contentMode: .fit)
-                        .padding(16)
+            VStack(spacing: 13) {
+                HStack {
+                    AvatarView(imageUrl: author.profilePhotoURL, size: 80)
+                        .padding(.trailing, 12)
+                    VStack(alignment: .leading, spacing: 6) {
+                        HStack {
+                            PlainText(author.safeName)
+                                .lineLimit(1)
+                                .font(.title3)
+                                .fontWeight(.bold)
+                                .foregroundColor(Color.primaryTxt)
+                                .multilineTextAlignment(.leading)
+                                .frame(maxWidth: .infinity, alignment: .leading)
+                            if author.muted {
+                                Text(Localized.muted.string)
+                                    .font(.subheadline)
+                                    .foregroundColor(Color.secondaryText)
+                            }
+                        }
+                        
+                        if !(author.uns ?? "").isEmpty {
+                            UNSNameView(author: author)
+                        } else {
+                            NIP05View(author: author)
+                        }
+                        
+                        if let bio = author.about {
+                            PlainText(bio)
+                                .foregroundColor(.secondaryText)
+                                .font(.claritySubheadline)
+                                .multilineTextAlignment(.leading)
+                                .lineSpacing(5)
+                                .lineLimit(2)
+                        }
+                    }
                 }
-                
-                Text(author.safeName)
-                    .padding(EdgeInsets(top: 9, leading: 10, bottom: 5, trailing: 10))
-                    .lineLimit(1)
-                    .foregroundColor(.primaryTxt)
-                    .font(.subheadline)
-                
-                if author.muted {
-                    Text(Localized.muted.string)
-                        .font(.subheadline)
-                        .foregroundColor(Color.secondaryText)
-                } else {
-                    Text(author.nip05 ?? author.npubString ?? "")
-                        .foregroundColor(.secondaryText)
-                        .font(.footnote)
-                        .lineLimit(1)
-                        .padding(EdgeInsets(top: 0, leading: 10, bottom: 5, trailing: 10))
-                    Text(author.about ?? "")
-                        .frame(maxWidth: .infinity, alignment: .center)
-                        .multilineTextAlignment(.center)
-                        .allowsHitTesting(false)
-                        .padding(EdgeInsets(top: 0, leading: 10, bottom: 0, trailing: 10))
-                }
-                Spacer(minLength: 9)
-                
-                if let first = knownFollowers[safe: 0]?.source {
-                    AuthorCardKnownFollowersView(
-                        first: first,
-                        knownFollowers: knownFollowers,
-                        followers: followers
-                    )
-                    Spacer(minLength: 9)
-                }
+                KnownFollowersView(author: author)
             }
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .aspectRatio(goldenRatio, contentMode: ContentMode.fill)
+            .padding(15)
             .background(
-                LinearGradient.cardBackground
+                LinearGradient(
+                    colors: [Color.cardBgTop, Color.cardBgBottom],
+                    startPoint: .top,
+                    endPoint: .bottom
+                )
             )
-            .cornerRadius(15)
-            .padding(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0))
+            .cornerRadius(cornerRadius)
         }
-        .buttonStyle(CardButtonStyle(style: .golden))
+        .buttonStyle(CardButtonStyle(style: .compact))
+        .listRowInsets(EdgeInsets())
     }
+
+    var cornerRadius: CGFloat { 15 }
 }
 
-struct AuthorCard_Previews: PreviewProvider {
-    static var previewData = PreviewData()
-    static var previews: some View {
-        StaggeredGrid(
-            list: [previewData.alice, previewData.bob, previewData.eve], 
-            columns: 2, 
-            content: { author in
-                AuthorCard(author: author)
-            }
-        ) 
+#Preview {
+    var previewData = PreviewData()
+    
+    return VStack {
+        AuthorCard(author: previewData.alice) 
+        AuthorCard(author: previewData.unsAuthor) 
+        AuthorCard(author: previewData.bob) 
     }
+    .padding()
+    .inject(previewData: previewData)
 }