+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fixed the 3d card effect on the Notifications tab.
+- Updated layout for search results and mention autocomplete cards.
 
 ## [0.1 (93)] - 2023-11-10Z